+++ cpp/src/barretenberg/stdlib/primitives/composers/composers.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "barretenberg/honk/composer/standard_plonk_composer.hpp"
+#include "barretenberg/plonk/composer/splitting_tmp/standard_plonk_composer.hpp"
 #include "barretenberg/honk/composer/standard_honk_composer.hpp"
 #include "barretenberg/plonk/composer/standard_composer.hpp"
 #include "barretenberg/plonk/composer/turbo_composer.hpp"
