+++ cpp/src/barretenberg/stdlib/primitives/field/field.hpp
@@ -3,7 +3,7 @@
 #include "../composers/composers_fwd.hpp"
 #include "../witness/witness.hpp"
 #include "barretenberg/honk/composer/standard_honk_composer.hpp"
-#include "barretenberg/honk/composer/standard_plonk_composer.hpp"
+#include "barretenberg/plonk/composer/splitting_tmp/standard_plonk_composer.hpp"
 #include "barretenberg/common/assert.hpp"
 
 namespace plonk {
